mcp-playwright-automation/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-playwright-automation"
version = "0.1.0"
description = "Model Context Protocol server for Playwright web automation in WhatsApp campaign systems"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "FullForce Academia", email = "dev@fullforce.com" }
]
keywords = ["mcp", "playwright", "automation", "whatsapp", "n8n", "google-sheets"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "asyncio",
    "httpx>=0.24.0",
    "structlog>=23.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "jsonschema>=4.17.0",
    "pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0"
]

[project.scripts]
mcp-playwright-automation = "mcp_playwright_automation.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_playwright_automation"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"